fix some narrowing warnings

+++ src/Algorithms/AlgorithmCiftiCorrelationGradient.cxx
@@ -793,12 +793,12 @@ void AlgorithmCiftiCorrelationGradient::processVolumeComponent(StructureEnum::En
     int64_t offset[3];
     if (mapSize > 0)
     {//make a voxel bounding box to minimize memory usage
-        int extrema[6] = { myMap[0].m_ijk[0],
-            myMap[0].m_ijk[0],
-            myMap[0].m_ijk[1],
-            myMap[0].m_ijk[1],
-            myMap[0].m_ijk[2],
-            myMap[0].m_ijk[2]
+        int64_t extrema[6] = { myMap[0].m_ijk[0],
+                               myMap[0].m_ijk[0],
+                               myMap[0].m_ijk[1],
+                               myMap[0].m_ijk[1],
+                               myMap[0].m_ijk[2],
+                               myMap[0].m_ijk[2]
         };
         for (int64_t i = 1; i < mapSize; ++i)
         {
@@ -942,12 +942,12 @@ void AlgorithmCiftiCorrelationGradient::processVolumeComponent(StructureEnum::En
     int64_t offset[3];
     if (mapSize > 0)
     {//make a voxel bounding box to minimize memory usage
-        int extrema[6] = { myMap[0].m_ijk[0],
-            myMap[0].m_ijk[0],
-            myMap[0].m_ijk[1],
-            myMap[0].m_ijk[1],
-            myMap[0].m_ijk[2],
-            myMap[0].m_ijk[2]
+        int64_t extrema[6] = { myMap[0].m_ijk[0],
+                               myMap[0].m_ijk[0],
+                               myMap[0].m_ijk[1],
+                               myMap[0].m_ijk[1],
+                               myMap[0].m_ijk[2],
+                               myMap[0].m_ijk[2]
         };
         for (int64_t i = 1; i < mapSize; ++i)
         {
